MEPA-1231: Create private functions and make lan80xx_mcu.c as public APIs with proper MEPA lock/unlock.
